requirements.txt
streamlit
ctransformers
sqlalchemy
pydantic
pyyaml
python-dotenv
pandas
groq
